+++ ledger/tests/test_api/test_char_journal_api.py
@@ -9,13 +9,9 @@ from app_utils.testing import create_user_from_evecharacter
 from ledger import app_settings
 from ledger.api.character.journal import LedgerJournalApiEndpoints
 from ledger.api.schema import Character
+from ledger.models.characteraudit import CharacterWalletJournalEntry
 from ledger.tests.testdata.load_allianceauth import load_allianceauth
 from ledger.tests.testdata.load_ledger import load_ledger_all
-
-if app_settings.LEDGER_MEMBERAUDIT_USE:
-    from memberaudit.models import CharacterWalletJournalEntry
-else:
-    from ledger.models.characteraudit import CharacterWalletJournalEntry
 
 
 class ManageApiJournalCharEndpointsTest(TestCase):
@@ -62,20 +58,12 @@ class ManageApiJournalCharEndpointsTest(TestCase):
                     "id": journal.entry_id,
                     "date": journal.date.strftime("%Y-%m-%dT%H:%M:%SZ"),
                     "first_party": {
-                        "id": (
-                            journal.first_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.first_party.id
-                        ),
+                        "id": (journal.first_party.eve_id),
                         "name": journal.first_party.name,
                         "cat": journal.first_party.category,
                     },
                     "second_party": {
-                        "id": (
-                            journal.second_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.second_party.id
-                        ),
+                        "id": (journal.second_party.eve_id),
                         "name": journal.second_party.name,
                         "cat": journal.second_party.category,
                     },
@@ -108,20 +96,12 @@ class ManageApiJournalCharEndpointsTest(TestCase):
                     "id": journal.entry_id,
                     "date": journal.date.strftime("%Y-%m-%dT%H:%M:%SZ"),
                     "first_party": {
-                        "id": (
-                            journal.first_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.first_party.id
-                        ),
+                        "id": (journal.first_party.eve_id),
                         "name": journal.first_party.name,
                         "cat": journal.first_party.category,
                     },
                     "second_party": {
-                        "id": (
-                            journal.second_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.second_party.id
-                        ),
+                        "id": (journal.second_party.eve_id),
                         "name": journal.second_party.name,
                         "cat": journal.second_party.category,
                     },
@@ -154,20 +134,12 @@ class ManageApiJournalCharEndpointsTest(TestCase):
                     "id": journal.entry_id,
                     "date": journal.date.strftime("%Y-%m-%dT%H:%M:%SZ"),
                     "first_party": {
-                        "id": (
-                            journal.first_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.first_party.id
-                        ),
+                        "id": (journal.first_party.eve_id),
                         "name": journal.first_party.name,
                         "cat": journal.first_party.category,
                     },
                     "second_party": {
-                        "id": (
-                            journal.second_party.eve_id
-                            if not app_settings.LEDGER_MEMBERAUDIT_USE
-                            else journal.second_party.id
-                        ),
+                        "id": (journal.second_party.eve_id),
                         "name": journal.second_party.name,
                         "cat": journal.second_party.category,
                     },