+++ ledger/tests/test_hooks.py
@@ -4,11 +4,7 @@ from unittest.mock import patch
 from django.test import TestCase
 
 from ledger.errors import LedgerImportError
-from ledger.hooks import (
-    get_corp_models_and_string,
-    get_extension_logger,
-    get_models_and_string,
-)
+from ledger.hooks import get_corp_models_and_string, get_extension_logger
 
 MODULE_PATH = "ledger.hooks"
 
@@ -27,7 +23,6 @@ class TestApiHelperCorpStatsImport(TestCase):
         sys.modules = self.original_sys_modules
 
     @patch(MODULE_PATH + ".app_settings.LEDGER_CORPSTATS_TWO", True)
-    @patch(MODULE_PATH + ".app_settings.LEDGER_MEMBERAUDIT_USE", True)
     def test_packages_are_not_installed(self):
         with patch.dict(
             sys.modules,
@@ -35,12 +30,3 @@ class TestApiHelperCorpStatsImport(TestCase):
         ):
             with self.assertRaises(LedgerImportError):
                 _ = get_corp_models_and_string()
-
-        with patch.dict(
-            sys.modules,
-            {k: None for k in list(sys.modules) if k.startswith("memberaudit")},
-        ):
-            with self.assertRaises(LedgerImportError):
-                CharacterMiningLedger, CharacterWalletJournalEntry = (
-                    get_models_and_string()
-                )